requirements.txt
# Django Core
Django>=5.0,<5.1

# Persian (Farsi) Localization
django-jalali>=6.0.1
django-jalali-date>=1.0.1

# Database
# SQLite is included with Python, but for production you might need:
# psycopg2-binary>=2.9.9  # PostgreSQL
# mysqlclient>=2.2.0      # MySQL

# Production Server (optional)
# gunicorn>=21.2.0
# whitenoise>=6.6.0
